app/requirements.txt
Django==3.2.23
environs==7.3.1
djangorestframework==3.14.0
